# Conflicts:
#	Modulo-Cuentas-Pll/Modulo-Cuentas-Pll-web/src/main/java/ec/edu/espe/arquitectura/rest/api/CuentaResource.java

+++ Modulo-Cuentas-Pll/Modulo-Cuentas-Pll-ejb/src/main/java/ec/edu/espe/arquitectura/model/Interes.java
@@ -43,7 +43,7 @@ public class Interes implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "ID_INTERES")
-    private BigDecimal idInteres;
+    private Integer idInteres;
     @Basic(optional = false)
     @NotNull
     @Column(name = "PORCENTAJE_INTERES")
@@ -65,22 +65,22 @@ public class Interes implements Serializable {
     public Interes() {
     }
 
-    public Interes(BigDecimal idInteres) {
+    public Interes(Integer idInteres) {
         this.idInteres = idInteres;
     }
 
-    public Interes(BigDecimal idInteres, BigDecimal porcentajeInteres, BigDecimal valorMin, BigDecimal valorMax) {
+    public Interes(Integer idInteres, BigDecimal porcentajeInteres, BigDecimal valorMin, BigDecimal valorMax) {
         this.idInteres = idInteres;
         this.porcentajeInteres = porcentajeInteres;
         this.valorMin = valorMin;
         this.valorMax = valorMax;
     }
 
-    public BigDecimal getIdInteres() {
+    public Integer getIdInteres() {
         return idInteres;
     }
 
-    public void setIdInteres(BigDecimal idInteres) {
+    public void setIdInteres(Integer idInteres) {
         this.idInteres = idInteres;
     }
